// mpl-token-metadata/src/generated/types/collection.ts
/**
 * This code was AUTOGENERATED using the kinobi library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun kinobi to update it.
 *
 * @see https://github.com/kinobi-so/kinobi
 */

import {
  combineCodec,
  getAddressDecoder,
  getAddressEncoder,
  getBooleanDecoder,
  getBooleanEncoder,
  getStructDecoder,
  getStructEncoder,
  type Address,
  type Codec,
  type Decoder,
  type Encoder,
} from '@solana/web3.js';

export type Collection = { verified: boolean; key: Address };

export type CollectionArgs = Collection;

export function getCollectionEncoder(): Encoder<CollectionArgs> {
  return getStructEncoder([
    ['verified', getBooleanEncoder()],
    ['key', getAddressEncoder()],
  ]);
}

export function getCollectionDecoder(): Decoder<Collection> {
  return getStructDecoder([
    ['verified', getBooleanDecoder()],
    ['key', getAddressDecoder()],
  ]);
}

export function getCollectionCodec(): Codec<CollectionArgs, Collection> {
  return combineCodec(getCollectionEncoder(), getCollectionDecoder());
}